requirements.txt
flask
psycopg2-binary
redis
redis-server
python-dotenv
gunicorn